welp we can't really nail down the codecs after all
it's a fucking pain.
you'd expect it would be like, avformat supports
containers A, B, and C,
avcodec supports codecs X, Y, and Z,
and maybe A only allows X, Y, etc.
but in fact only a handful of specific combinations
of [container format]+[audio codec]+[video codec] are supported.